apps/bot/requirements.txt
# httpx совместим с python-telegram-bot 20.x (см. зависимости PTB)
python-telegram-bot[job-queue]==20.7
httpx==0.25.2
python-dotenv>=1.0.0
pytz>=2024.1
